amazon-reviews/scripts/src/requirements.txt
absl-py
aiodns
aiohttp
aiohttp-socks
aiosignal
altair
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
astunparse
async-generator
async-timeout
attrs
autopep8
backcall
beautifulsoup4
billiard
bleach
blinker
bs4
build
cachetools
catboost
category-encoders
certifi
cffi
chardet
charset-normalizer
cheap-repr
click
cloudpickle
cmake
coala-utils
colorama
comm
commonmark
contourpy
cryptography
cycler
dash
dataclasses
debugpy
decorator
defusedxml
distlib
distro
docopt
docutils
elastic-transport
elasticsearch
entrypoints
executing
fake-useragent
Faker
fastjsonschema
feather-format
filelock
Flask
flatbuffers
fonttools
fqdn
freeze-requirements
frozenlist
fsspec
gast
geographiclib
geopy
gitdb
GitPython
google-auth
google-auth-oauthlib
google-pasta
googletransx
graphviz
grpcio
h11
h5py
huggingface-hub
icecream
idna
imageio
imageio-ffmpeg
imbalanced-learn
imblearn
importlib-metadata
ipykernel
ipython
ipython-genutils
ipywidgets
iso3166
isoduration
itsdangerous
jedi
Jinja2
joblib
jsonpointer
jsonschema
jupyterlab-widgets
keras
kiwisolver
lazy_loader
libclang
lightgbm
lit
llvmlite
locket
lux
lux-api
lux-widget
lxml
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
MouseInfo
moviepy
mozdownload
mozfile
mozinfo
mpmath
multidict
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nltk
notebook
notebook_shim
numba
numpy
oauthlib
opencv-contrib-python
opencv-python
opt-einsum
outcome
packaging
pandas
pandocfilters
parso
partd
patsy
pdfminer.six
pdfplumber
pep517
pexpect
pickleshare
pil-video
Pillow
pipreqs
platformdirs
platinfo
plotly
polars
portpicker
proglog
progress
progressbar2
prometheus-client
prompt-toolkit
protobuf
psutil
ptyprocess
pure-eval
py-firefox-driver-manager
pyarrow
pyasn1
pyasn1-modules
PyAutoGUI
pycares
pycodestyle
pycparser
pydeck
PyGetWindow
Pygments
pyjokes
Pympler
PyMsgBox
pyOpenSSL
pyparsing
pyperclip
PyRect
pyrsistent
PyScreeze
PySocks
pysqlite3
pytweening
pytz
pytz-deprecation-shim
PyWavelets
pywhatkit
pywin32-ctypes
PyYAML
pyzmq
redo
regex
requests
requests-oauthlib
rich
rich-cli
rich-rst
rsa
schedule
scikit-image
scikit-learn
scipy
seaborn
selenium
semver
Send2Trash
sh
shap
shapash
simplegeneric
simplemma
six
slicer
smmap
sniffio
snoop
snscrape
sortedcontainers
soupsieve
SpeechRecognition
stack-data
statsmodels
streamlit
style
sympy
tenacity
tensorboard
terminado
textual
threadpoolctl
tifffile
tinycss2
tokenizers
toml
tomli
toolz
tornado
tqdm
traitlets
transformers
treeherder-client
trio
trio-websocket
triton
tweepy
twitter-scraper-without-api
twitterscraper
typing_extensions
tzdata
tzlocal
update
uri-template
urllib3
validators
virtualenv
Wand
watchdog
wcwidth
webcolors
webencodings
websocket-client
Werkzeug
widgetsnbextension
wikipedia
wordcloud
wrapt
wsproto
xgboost
yarg
yarl
zipp